AI_Service/requirements.txt
fastapi
uvicorn
numpy
opencv-python-headless
insightface
onnxruntime
faiss-cpu
python-multipart
requests
